requirements.txt
groq
duckduckgo-search
newsapi-python
yfinance
textblob
pandas
streamlit
plotly
python-dotenv
numpy
pdfplumber
pandas-ta
